Customer service updated

+++ customer-service/src/models/user.ts
@@ -4,7 +4,7 @@ const User = (db: Sequelize) => {
   const user = db.define('user', {
     name: DataTypes.STRING,
     email: DataTypes.STRING,
-    balance: DataTypes.DOUBLE,
+    balance: DataTypes.DECIMAL,
   });
   return user;
 };
